+++ Magic/Spells/MeteorSpell.cs
@@ -73,7 +73,7 @@ namespace Magic.Spells
                 {
                     for (int ix = -i; ix <= i; ++ix)
                         for (int iy = -i; iy <= i; ++iy)
-                            Game1.createRadialDebris(this.loc, Game1.objectSpriteSheetName, new Rectangle(352, 400, 32, 32), 4, (int)this.position.X + ix * 20, (int)this.position.Y + iy * 20, 15 - 14 + Meteor.rand.Next(15 - 14), (int)((double)this.position.Y / (double)Game1.tileSize) + 1, new Color(255, 255, 255, 255), 4.0f);
+                            Game1.createRadialDebris(this.loc, Game1.objectSpriteSheetName, new Rectangle(352, 400, 32, 32), 4, (int)this.position.X + ix * 20, (int)this.position.Y + iy * 20, 15 - 14 + Meteor.rand.Next(15 - 14), (int)(this.position.Y / (double)Game1.tileSize) + 1, new Color(255, 255, 255, 255), 4.0f);
                 }
                 foreach (var npc in this.source.currentLocation.characters)
                 {
@@ -98,7 +98,7 @@ namespace Magic.Spells
         public void Draw(SpriteBatch spriteBatch)
         {
             Vector2 drawPos = Game1.GlobalToLocal(new Vector2(this.position.X, this.position.Y - this.height));
-            spriteBatch.Draw(Game1.objectSpriteSheet, drawPos, new Rectangle(352, 400, 32, 32), Color.White, 0, new Vector2(16, 16), 2 + 8, SpriteEffects.None, (float)(((double)this.position.Y - this.height + (double)(Game1.tileSize * 3 / 2)) / 10000.0));
+            spriteBatch.Draw(Game1.objectSpriteSheet, drawPos, new Rectangle(352, 400, 32, 32), Color.White, 0, new Vector2(16, 16), 2 + 8, SpriteEffects.None, (float)(((double)this.position.Y - this.height + Game1.tileSize * 3 / 2) / 10000.0));
         }
     }
 }
